Dining cryptographers and philosophers problems
1. Add dining cryptographers and philosophers problems.
  a. example*/DiningCrypto.prot
  a. example*/DiningPhilo.prot
  a. example*/DiningPhiloRand.prot

2. Allow actions that aren't self-disabling if the self-disabling
   version violates safety.
  c. synthesis.cc/hh  coexist_ck()
  a. xnsys.cc/hh  Net::safe_ck() PcSymm::init_representative()

3. Check to ensure synchronous semantics don't break closure of assumed
   states during synthesis.
  a. synthesis.cc  revise_actions_alone()

4. Fix synthesis for the case where all states are invariant.
  b. synthesis.cc  RankDeadlocksMRV() SynthesisCtx::add()

+++ s/runudp.sh
@@ -1,20 +1,33 @@
 #!/bin/sh
 
-npcs=50
+npcs=$1
+
+if [ ! "$npcs" ]
+then
+  echo "Give a number of processes!"
+  exit 1
+fi
+
+ExeName=$(readlink -f "$0")
+udp=$(dirname $(dirname "$ExeName"))/src/udp
 
 rm -f udp-host.*
 rm -f out.log.*
+rm -f pids
 
 for i in $(seq 0 $(expr $npcs - 1))
 do
-  ./udp $i $npcs out.log.$i 2>&1 &
-  #valgrind --track-origins=yes ./udp $i $npcs out.log.$i 2>&1 &
-  #./udp $i $npcs 2>&1 | tee out.log.$i &
-  #./udp $i $npcs 2>&1 &
+  "$udp" $i $npcs out.log.$i 2>&1 &
+  echo $! >> pids
+  #valgrind --track-origins=yes "$udp" $i $npcs out.log.$i 2>&1 &
+  #"$udp" $i $npcs 2>&1 | tee out.log.$i &
+  #"$udp" $i $npcs 2>&1 &
 done
 
 for i in $(seq 0 $(expr $npcs - 1))
 do
   wait
 done
+
+rm -f pids
